+++ api/audit/tasks.py
@@ -3,7 +3,10 @@ import typing
 
 from django.contrib.auth import get_user_model
 
-from audit.constants import FEATURE_STATE_WENT_LIVE_MESSAGE
+from audit.constants import (
+    FEATURE_STATE_UPDATED_BY_CHANGE_REQUEST_MESSAGE,
+    FEATURE_STATE_WENT_LIVE_MESSAGE,
+)
 from audit.models import AuditLog, RelatedObjectType
 from task_processor.decorators import register_task_handler
 
@@ -12,6 +15,21 @@ logger = logging.getLogger(__name__)
 
 @register_task_handler()
 def create_feature_state_went_live_audit_log(feature_state_id: int):
+    _create_feature_state_audit_log_for_change_request(
+        feature_state_id, FEATURE_STATE_WENT_LIVE_MESSAGE
+    )
+
+
+@register_task_handler()
+def create_feature_state_updated_by_change_request_audit_log(feature_state_id: int):
+    _create_feature_state_audit_log_for_change_request(
+        feature_state_id, FEATURE_STATE_UPDATED_BY_CHANGE_REQUEST_MESSAGE
+    )
+
+
+def _create_feature_state_audit_log_for_change_request(
+    feature_state_id: int, msg_template: str
+):
     from features.models import FeatureState
 
     feature_state = FeatureState.objects.filter(id=feature_state_id).first()
@@ -25,7 +43,7 @@ def create_feature_state_went_live_audit_log(feature_state_id: int):
     if not feature_state.change_request:
         raise RuntimeError("Feature state must have a change request")
 
-    message = FEATURE_STATE_WENT_LIVE_MESSAGE % (
+    log = msg_template % (
         feature_state.feature.name,
         feature_state.change_request.title,
     )
@@ -34,7 +52,7 @@ def create_feature_state_went_live_audit_log(feature_state_id: int):
         related_object_type=RelatedObjectType.FEATURE_STATE.name,
         environment=feature_state.environment,
         project=feature_state.environment.project,
-        log=message,
+        log=log,
         is_system_event=True,
     )
