+++ api/tests/unit/audit/test_unit_audit_tasks.py
@@ -1,7 +1,13 @@
+from audit.constants import (
+    FEATURE_STATE_UPDATED_BY_CHANGE_REQUEST_MESSAGE,
+    FEATURE_STATE_WENT_LIVE_MESSAGE,
+)
 from audit.models import AuditLog
 from audit.related_object_type import RelatedObjectType
 from audit.tasks import (
     create_audit_log_from_historical_record,
+    create_feature_state_updated_by_change_request_audit_log,
+    create_feature_state_went_live_audit_log,
     create_segment_priorities_changed_audit_log,
 )
 from features.models import FeatureSegment
@@ -183,3 +189,87 @@ def test_create_segment_priorities_changed_audit_log(
         environment=environment,
         log=f"Segment overrides re-ordered for feature '{feature.name}'.",
     ).exists()
+
+
+def test_create_feature_state_went_live_audit_log(change_request_feature_state):
+    # Given
+    message = FEATURE_STATE_WENT_LIVE_MESSAGE % (
+        change_request_feature_state.feature.name,
+        change_request_feature_state.change_request.title,
+    )
+    feature_state_id = change_request_feature_state.id
+
+    # When
+    create_feature_state_went_live_audit_log(feature_state_id)
+
+    # Then
+    assert (
+        AuditLog.objects.filter(
+            related_object_id=feature_state_id, is_system_event=True, log=message
+        ).count()
+        == 1
+    )
+
+
+def test_create_feature_state_updated_by_change_request_audit_log(
+    change_request_feature_state,
+):
+    # Given
+    message = FEATURE_STATE_UPDATED_BY_CHANGE_REQUEST_MESSAGE % (
+        change_request_feature_state.feature.name,
+        change_request_feature_state.change_request.title,
+    )
+    feature_state_id = change_request_feature_state.id
+
+    # When
+    create_feature_state_updated_by_change_request_audit_log(feature_state_id)
+
+    # Then
+    assert (
+        AuditLog.objects.filter(
+            related_object_id=feature_state_id, is_system_event=True, log=message
+        ).count()
+        == 1
+    )
+
+
+def test_create_feature_state_updated_by_change_request_audit_log_does_nothing_if_feature_state_deleted(
+    change_request_feature_state,
+):
+    # Given
+    change_request_feature_state.delete()
+    feature_state_id = change_request_feature_state.id
+
+    # When
+    create_feature_state_went_live_audit_log(feature_state_id)
+
+    # Then
+    assert (
+        AuditLog.objects.filter(
+            related_object_id=feature_state_id, is_system_event=True
+        ).count()
+        == 0
+    )
+
+
+def test_create_feature_state_wen_live_audit_log_does_nothing_if_feature_state_deleted(
+    change_request_feature_state,
+):
+    # Given
+    message = FEATURE_STATE_WENT_LIVE_MESSAGE % (
+        change_request_feature_state.feature.name,
+        change_request_feature_state.change_request.title,
+    )
+    change_request_feature_state.delete()
+    feature_state_id = change_request_feature_state.id
+
+    # When
+    create_feature_state_went_live_audit_log(feature_state_id)
+
+    # Then
+    assert (
+        AuditLog.objects.filter(
+            related_object_id=feature_state_id, is_system_event=True, log=message
+        ).count()
+        == 0
+    )
